+++ schema/readout/bufferedfilestreamer.jsonnet
