+++ test/fakereadout-commands.json
@@ -26,7 +26,7 @@
                         ]
                     },
                     "inst": "snb-writer",
-                    "plugin": "SNBWriter"
+                    "plugin": "BufferedFileStreamer"
                 },
                 {
                     "data": {
